requirements.txt
langchain>=0.0.325
langgraph>=0.0.10
fastapi>=0.104.0
uvicorn>=0.23.2
redis>=5.0.1
python-dotenv>=1.0.0
openai>=1.2.3
requests>=2.31.0
python-multipart>=0.0.6
langsmith>=0.0.52
streamlit>=1.28.0
celery>=5.3.4
aiohttp>=3.8.6
pytest>=7.4.3
pytest-asyncio>=0.21.1
pydantic>=2.4.2
black>=23.10.1
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.1
docker>=6.1.3
ddgs